Feature/edit subscriptions

+++ backend/src/main/java/ftn/kts/controller/SubscriptionController.java
@@ -37,8 +37,8 @@ public class SubscriptionController {
 	}
 
 	@GetMapping
-	@PreAuthorize("hasAnyRole('USER', 'ADMIN')")
-	public ResponseEntity<Page<SubscriptionDTO>> getAllSubscriptions(@RequestParam(defaultValue = "0") Integer pageNo,
+	@PreAuthorize("hasAnyRole('USER')")
+	public ResponseEntity<Page<SubscriptionDTO>> getMySubscriptions(@RequestParam(defaultValue = "0") Integer pageNo,
 			@RequestParam(defaultValue = "10") Integer pageSize, @RequestParam(defaultValue = "id") String sortBy,
 			@RequestParam(defaultValue = "true") String descending) {
 		Pageable paging;
@@ -46,8 +46,11 @@ public class SubscriptionController {
 			paging = PageRequest.of(pageNo, pageSize, Sort.by(Sort.Direction.DESC, sortBy));
 		else
 			paging = PageRequest.of(pageNo, pageSize, Sort.by(Sort.Direction.ASC, sortBy));
-		Page<SubscriptionDTO> subscriptions = service.getAllDTO(paging);
-		return new ResponseEntity<>(subscriptions, HttpStatus.OK);
+		Authentication currentUser = SecurityContextHolder.getContext().getAuthentication();
+		String username = currentUser.getName();
+		Page<SubscriptionDTO> subscriptions = service.getAllDTOForUsername(username, paging);
+		ResponseEntity<Page<SubscriptionDTO>> responseEntity = new ResponseEntity<>(subscriptions, HttpStatus.OK);
+		return responseEntity;
 	}
 
 	@GetMapping("/{id}")
@@ -140,7 +143,7 @@ public class SubscriptionController {
 	@PreAuthorize("hasRole('USER')")
 	public ResponseEntity<String> addSubscription(@Valid @RequestBody SubscriptionDTO dto) {
 		service.create(dto);
-		return new ResponseEntity<>("Successfully added subscription!", HttpStatus.CREATED);
+		return new ResponseEntity<>("{ \"message\": \"" + "Successfully added subscription!" + "}", HttpStatus.CREATED);
 	}
 
 	@PutMapping("/{id}")
@@ -157,8 +160,12 @@ public class SubscriptionController {
 
 	@DeleteMapping("/{id}")
 	@PreAuthorize("hasRole('USER')")
-	public ResponseEntity<String> deleteSubscription(@PathVariable("id") long id) {
+	public ResponseEntity<SubscriptionDTO> deleteSubscription(@PathVariable("id") long id) {
+		SubscriptionDTO dto = service.getOneDTO(id);
+		if (dto == null) {
+			return new ResponseEntity<SubscriptionDTO>(dto, HttpStatus.NOT_FOUND);
+		}
 		service.delete(id);
-		return new ResponseEntity<>("Successfully deleted subscription!", HttpStatus.OK);
+		return new ResponseEntity<>(dto, HttpStatus.OK);
 	}
 }
